+++ components/ListExpenses/ListExpenses.tsx
@@ -1,4 +1,4 @@
-
+import style from "./ListExpenses.module.css"
 
 export default function ListExpenses(){
