+++ components/SearchBar/SearchBar.tsx
@@ -1,4 +1,4 @@
-
+import style from "./SearchBar.module.css"
 
 export default function SearchBar (){
