+++ pages/index.js
@@ -1,6 +1,5 @@
 import Head from 'next/head'
-import Nav from "../components/Nav";
-import Footer from "../components/Footer";
+
 import Image from "next/image";
 import MainDescription from "../components/MainDescription";
 
@@ -12,13 +11,10 @@ export default function Home() {
                 <meta name="description" content="Generated by create next app"/>
                 <link rel="icon" href="/favicon.ico"/>
             </Head>
-            <Nav/>
             <main className="container h-full p-4 flex flex-1 flex-col items-center">
                 <MainDescription />
                 <p>Hola</p>
             </main>
-
-            <Footer/>
         </div>
     )
 }
