+++ pages/private/home/index.tsx
@@ -1,7 +1,7 @@
 import {useDispatch, useSelector} from "react-redux";
-import {AppState} from "../../redux";
+import {AppState} from "../../../redux";
 import {useRouter} from "next/router";
-import PrivateLayout from "../../layouts/PrivateLayout/PrivateLayout";
+
 
 export default function Home() {
     /// Selector para consultar el user
@@ -9,13 +9,12 @@ export default function Home() {
     const router = useRouter()
     /// Dispach para actualizar el user
     const dispatch = useDispatch()
-
+    
     if (!currentUser) {
         router.push('/login')
         return <></>
     }
     return (
-        <PrivateLayout>
             <div className="flex flex-col items-center">
                 <main className="container h-full p-4 flex flex-1 flex-col items-center justify-center">
                     <section>
@@ -23,6 +22,6 @@ export default function Home() {
                     </section>
                 </main>
             </div>
-        </PrivateLayout>
     )
 }
+
