+++ pages/private/search/search.module.css
@@ -1,0 +1,7 @@
+
+
+.Search{
+    display:flex;
+    flex-direction: column;
+    align-items: center;
+}
